// Kopy/src/app/modules/mantenimiento/models/productos-request.module.ts
export class ProductoRequest {
  id: number = 0;
  idCategoria: number = 0;
  idSubCategoria: number = 0;
  codigo: string ='';
  calificacion: number = 0;
  nombre: string ='';
  stock: number = 0;
  foto: string ='';
  marca: string ='';
  modelo: string ='';
  precioCompra: number = 0;
  precioVenta: number = 0;
  descripcion: string ='';
  estado: boolean =true;
}